refactor(aspect): Unify preprocessing mains
We split them in the beginning to omit the heavy boost::wave dependency
in case of the user not performing full preprocessing. Since, the final
implementation of the no preprocessing mode uses boost::wave as well, this
is no longer valid. At the same time having 2 main functions duplicates
code and in the past we already had inconsistent behavior due to that.
