+++ dwyu/aspect/dwyu.bzl
@@ -331,6 +331,7 @@ def _extract_includes_from_files(ctx, target, files, defines, cc_toolchain, attr
         # The source files could be a TreeArtifact! Thus, process each file as list, although we want to process the individual source files in parallel by default.
         args = make_param_file_args(ctx)
         args.add_all("--files", [file])
+        args.add("--mode", "fast" if ctx.attr._no_preprocessor else "full")
         args.add_all("--include_paths", include_paths)
         args.add_all("--system_include_paths", system_include_paths)
         if not ctx.attr._no_preprocessor:
