+++ dwyu/aspect/private/preprocessing/fast_parsing/BUILD
@@ -6,3 +6,20 @@ cc_library(
     hdrs = ["extract_includes.h"],
     visibility = ["//dwyu/aspect/private/preprocessing:__subpackages__"],
 )
+
+cc_library(
+    name = "process_files",
+    srcs = ["process_files.cpp"],
+    hdrs = ["process_files.h"],
+    implementation_deps = [
+        ":extract_includes",
+        "//dwyu/aspect/private/preprocessing/lib:included_file",
+        "//dwyu/private:utils",
+        "@boost.wave//:wave",
+        "@boost//:filesystem",
+    ],
+    visibility = ["//dwyu/aspect/private/preprocessing:__subpackages__"],
+    deps = [
+        "@nlohmann_json//:singleheader-json",
+    ],
+)
