+++ scripts/ensure_cpp11_compatibility.sh
@@ -9,7 +9,6 @@ set -o nounset
 TARGETS=(
     "//dwyu/aspect/private/analyze_includes:main"
     "//dwyu/aspect/private/preprocessing:main"
-    "//dwyu/aspect/private/preprocessing:main_no_preprocessing"
     "//dwyu/aspect/private/process_target:main_cc"
 )
